WUTTTT

+++ rsa.h
@@ -2,6 +2,6 @@
 #define _myfile_h
 
 int gcd(int a, int h);
-int rsa(unsigned long long hashKey);
+int rsa(int hashKey);
 
 #endif
